(fixup) Move defaultKafkaClientSupplier to under the internals
